Revert "Merge branch 'dev' into Feature/ShipModels-v01"
This reverts commit 85f77385cb07ba1bac24674989928b5f59a164d5, reversing
changes made to 7e3b777d98e6d89e5cc0866d6b5ab6edcce99a7a.

+++ TheFirmaTheBetter/Assets/Project/Scripts/ShipSelection/ShipParts/ShipResources.cs
@@ -92,7 +92,7 @@ namespace ShipParts
             Channels.OnPlayerStatsChanged?.Invoke(shipBuilder, shipStats);
         }
 
-        public float CurrentEnergyAmount => shipEnergy.CurrentEnergyAmount;
+        public int CurrentEnergyAmount => shipEnergy.CurrentEnergyAmount;
 
         public ShipStats ShipStats => shipStats;
